+++ packages/isaacscript-common/src/functions/bosses.ts
@@ -1,4 +1,8 @@
-import type { StageType } from "isaac-typescript-definitions";
+import type {
+  BossID,
+  LevelStage,
+  StageType,
+} from "isaac-typescript-definitions";
 import { EntityType, LokiVariant } from "isaac-typescript-definitions";
 import { VectorZero } from "../core/constants";
 import {
@@ -52,22 +56,20 @@ export function getAliveBosses(
 /**
  * Helper function to get the set of every boss in the game.
  *
- * The set contains strings with the entity type and variant, separated by a period.
- *
  * Note that this set does not include:
  * - mini-bosses (e.g. Ultra Pride, Krampus)
  * - bosses that do not appear in Boss Rooms (e.g. Uriel, Gabriel)
- * - the second phase of multi-phase bosses (e.g. Mega Satan 2), with the exception of The Beast
- *   (Ultra Famine, Ultra Pestilence, Ultra War, and Ultra Death are included)
+ * - the second phase of multi-phase bosses (e.g. Mega Satan 2, Ultra Famine, Ultra Pestilence,
+ *   Ultra War, Ultra Death)
  *
  * Also see the `getBossSet` and `getCombinedBossSet` functions.
  *
- * @param includeStoryBosses Optional. Whether to include "story" bosses like Mom and It Lives!
+ * @param includeStoryBosses Optional. Whether to include "story" bosses like Mom and It Lives.
  *                           Default is true.
  */
 export function getAllBossesSet(
   includeStoryBosses = true,
-): ReadonlySet<string> {
+): ReadonlySet<BossID> {
   return includeStoryBosses
     ? ALL_BOSSES_SET
     : ALL_BOSSES_EXCLUDING_STORY_BOSSES_SET;
@@ -77,25 +79,18 @@ export function getAllBossesSet(
  * Helper function to get the set of vanilla bosses for a particular stage and stage type
  * combination.
  *
- * The set contains strings with the entity type and variant, separated by a period.
- *
  * Also see the `getAllBossesSet` and `getCombinedBossSet` functions.
  */
 export function getBossSet(
-  stage: int,
+  stage: LevelStage,
   stageType: StageType,
-): ReadonlySet<string> | undefined {
+): ReadonlySet<BossID> | undefined {
   const stageTypeMap = STAGE_TO_STAGE_TYPE_TO_BOSS_SET_MAP.get(stage);
   if (stageTypeMap === undefined) {
     return undefined;
   }
 
-  const bossSet = stageTypeMap.get(stageType);
-  if (bossSet === undefined) {
-    return undefined;
-  }
-
-  return bossSet;
+  return stageTypeMap.get(stageType);
 }
 
 /**
@@ -121,22 +116,15 @@ export function getBosses(
 
 /**
  * Helper function to get the set of vanilla bosses for a particular stage across all of the stage
- * types. For example, specifying a stage of 2 will return a set with all of the bosses for
- * Basement, Cellar, Burning Basement, Downpour, and Dross.
- *
- * The set contains strings with the entity type and variant, separated by a period.
+ * types. For example, specifying `LevelStage.BASEMENT_2` will return a set with all of the bosses
+ * for Basement, Cellar, Burning Basement, Downpour, and Dross.
  *
  * Also see the `getAllBossesSet` and `getBossSet` functions.
  */
 export function getCombinedBossSet(
-  stage: int,
-): ReadonlySet<string> | undefined {
-  const bossSet = STAGE_TO_COMBINED_BOSS_SET_MAP.get(stage);
-  if (bossSet === undefined) {
-    return undefined;
-  }
-
-  return bossSet;
+  stage: LevelStage,
+): ReadonlySet<BossID> | undefined {
+  return STAGE_TO_COMBINED_BOSS_SET_MAP.get(stage);
 }
 
 /** Helper function to check if the provided NPC is a Sin miniboss, such as Sloth or Lust. */